+++ instrumentation/github.com/gin-gonic/gin/otelgin/gin_test.go
@@ -493,12 +493,43 @@ func TestTemporaryFormFileRemove(t *testing.T) {
 
 	router := gin.New()
 	router.MaxMultipartMemory = 1
+
+	// We register three middlewares here, with the otel one in the middle.
+	// When the response is returned, the post-handler execution order is:
+	// MiddlewareC -> otel -> MiddlewareA (reverse of registration order).
+	// In MiddlewareC, the multipart form file still exists, so opening it succeeds.
+	// However, by the time MiddlewareA runs, the otel middleware has already
+	// removed the temporary multipart files, so opening the file results in a
+	// "file does not exist" error.
+
+	// MiddlewareA
+	router.Use(func(c *gin.Context) {
+		c.Next()
+		form, _ := c.MultipartForm()
+		files := form.File["files"]
+		require.Len(t, files, 1)
+		_, err := files[0].Open()
+		require.ErrorIs(t, err, fs.ErrNotExist)
+
+	})
+
 	router.Use(otelgin.Middleware("foobar", otelgin.WithTracerProvider(provider)))
+
+	// MiddlewareC
+	router.Use(func(c *gin.Context) {
+		c.Next()
+		form, _ := c.MultipartForm()
+		files := form.File["files"]
+		require.Len(t, files, 1)
+		_, err := files[0].Open()
+		require.NoError(t, err)
+	})
+
 	var fileHeader *multipart.FileHeader
 	router.POST("/upload", func(c *gin.Context) {
-		_, err := c.FormFile("file")
+		_, err := c.FormFile("files")
 		require.NoError(t, err)
-		fileHeader = c.Request.MultipartForm.File["file"][0]
+		fileHeader = c.Request.MultipartForm.File["files"][0]
 		_, err = fileHeader.Open()
 		require.NoError(t, err)
 		c.JSON(http.StatusOK, nil)
@@ -507,7 +538,7 @@ func TestTemporaryFormFileRemove(t *testing.T) {
 	var body bytes.Buffer
 
 	mw := multipart.NewWriter(&body)
-	fw, err := mw.CreateFormFile("file", "file")
+	fw, err := mw.CreateFormFile("files", "file")
 	require.NoError(t, err)
 
 	_, err = fw.Write([]byte("hello world"))